Move timer functionality to FinishManager
