+++ Assets/Prefabs/FinishLine.prefab
@@ -10,6 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 3428338866642140752}
   - component: {fileID: 1757832134004540595}
+  - component: {fileID: 8879859643127342890}
   m_Layer: 0
   m_Name: FinishLine
   m_TagString: Finish
@@ -54,6 +55,21 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 0.65600586, y: 0.16919388, z: 9.96092}
   m_Center: {x: 0.2617798, y: 0, z: 0.033745766}
+--- !u!114 &8879859643127342890
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1481397041054031996}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 95c3abf96f35459478d8dd1c27562fd4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  BackToMainButton: {fileID: 0}
+  endLevelMenu: {fileID: 0}
+  currentTime: {fileID: 0}
 --- !u!1001 &3167708763923026423
 PrefabInstance:
   m_ObjectHideFlags: 0
