+++ Assets/Scripts/FinishManager.cs
@@ -1,3 +1,5 @@
+using System;
+using TMPro;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
@@ -7,23 +9,58 @@ public class FinishManager : MonoBehaviour
     [SerializeField] private Button BackToMainButton;
     [SerializeField] private GameObject endLevelMenu;
 
+    // for the level time mechanics
+    private bool timeStarted = false;
+    private float startTime;
+    [SerializeField] private TextMeshProUGUI currentTime;
+
     private void Awake()
     {
         BackToMainButton.onClick.AddListener(BackToMain);
         DisableEndLevelMenu();
     }
 
+    void Update()
+    {
+        if (Input.GetKey(KeyCode.Space)) {
+            timeStarted = true;
+            startTime = Time.time;
+        }
+
+        if (timeStarted) {
+            var guiTime = Time.time - startTime;
+
+            int minutes = (int)(guiTime / 60);
+            int seconds = (int)(guiTime % 60);
+            int fraction = (int)((guiTime * 100) % 100);
+
+            String time = String.Format ("{0:00}:{1:00}:{2:00}", minutes, seconds, fraction);
+            currentTime.SetText(time);
+        } 
+    }
+
     private void OnTriggerEnter(Collider other)
     {
         if(other.CompareTag("Player"))
         {
+            timeStarted = false;
+
+            var guiTime = Time.time - startTime;
+
+            int minutes = (int)(guiTime / 60);
+            int seconds = (int)(guiTime % 60);
+            int fraction = (int)((guiTime * 100) % 100);
+
+            String time = String.Format ("{0:00}:{1:00}:{2:00}", minutes, seconds, fraction);
+            currentTime.SetText(time);
+
             EnableEndLevelMenu();
         }
     }
 
     private void EnableEndLevelMenu()
     {
-        Time.timeScale = 0f; // save time somewher and display it
+        Time.timeScale = 0f; // save time somewhere and display it
         endLevelMenu.SetActive(true);
         Cursor.lockState = CursorLockMode.None;
         Cursor.visible = true;
