--- Assets/Scripts/LevelTimer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 39bb1361bf1918e4dbc048fa5425ca73
